% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotateDatasetEntities}
\alias{annotateDatasetEntities}
\title{Annotate the Entities in a dataset}
\usage{
annotateDatasetEntities(dataset, object, ...)
}
\arguments{
\item{dataset}{An object containing a dataset (e.g., an \code{eSet}).}

\item{object}{An object containing information to be used to annotate the Entities
(instances) of the dataset.}

\item{\dots}{Additional arguments to methods of \code{annotateDatasetEntities}.}
}
\value{
Each method returns an S4 object of the same class as \code{dataset}.
}
\description{
This is a generic function for annotating the Entities (instances) in a
dataset (e.g., the samples of an \code{eSet}) using some object.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
