% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotateDatasetFeatures}
\alias{annotateDatasetFeatures}
\title{Annotate the Features in a dataset}
\usage{
annotateDatasetFeatures(dataset, database, ...)
}
\arguments{
\item{dataset}{An object containing a dataset (e.g., an @code{eSet}).}

\item{database}{An object (e.g., an annotation database) containing information to be used
to annotate the Features (observations) of the dataset.}

\item{\dots}{Additional arguments to methods of @code{annotateDatasetFeatures}.}
}
\value{
Each method returns an S4 object of the same class as @code{dataset}.
}
\description{
This is a generic function for annotating the Features (observations) in a
dataset (e.g., the features of an \code{eSet}) using some object
(e.g., an annotation database).
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
