% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belongsToFeatureSpace.R
\name{belongsToFeatureSpace}
\alias{belongsToFeatureSpace}
\title{Tests whether each element of a vector belongs to a FeatureSpace}
\usage{
belongsToFeatureSpace(x, featureSpace)
}
\arguments{
\item{x}{A character vector}

\item{featureSpace}{A \code{\linkS4class{hiveFeatureSpaceEntity}} object}
}
\value{
A logical vector specifying whether each element of \code{x} belongs to the
given \code{FeatureSpace}
}
\description{
This convenience function tests whether each element of a character vector
belongs to a given \code{FeatureSpace}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
