% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwiseParity.R
\name{bitwiseParity}
\alias{bitwiseParity}
\alias{bitwiseParity,integer-method}
\alias{bitwiseParity,logical-method}
\alias{bitwiseParity,raw-method}
\alias{bitwiseParity,UUIDList-method}
\alias{bitwiseParity,list-method}
\title{Compute bitwise parity over a vector or list}
\usage{
bitwiseParity(x, even = TRUE)

\S4method{bitwiseParity}{integer}(x, even = TRUE)

\S4method{bitwiseParity}{logical}(x, even = TRUE)

\S4method{bitwiseParity}{raw}(x, even = TRUE)

\S4method{bitwiseParity}{UUIDList}(x, even = TRUE)

\S4method{bitwiseParity}{list}(x, even = TRUE)
}
\arguments{
\item{x}{An integer, logical or raw vector or array, or a character vector or array
containing hexadecimal values (leading \code{"0x"} optional)}

\item{even}{A logical value specifying whether the parity should be even or odd}
}
\value{
\itemize{
  \item{
    If \code{x} is of length zero, the output is a vector of length zero of
    the same storage mode as \code{x}.
  }
  \item{
    If \code{x} is a character vector consisting solely of empty strings, the
    output is \code{""}.
  }
  \item{
    If \code{x} is a character vector containing invalid hexadecimal
    strings, an NA value will be returned.
  }
  \item{
    Otherwise, a vector of length 1 of the same storage mode as \code{x} is
    returned, containing the bitwise parity computed across all elements of
    \code{x}.
  }
  \item{
    If the storage mode of \code{x} is \code{"character"}, the output will
    be in hexadecimal notation with no leading \code{"0x"}.
  }
}
}
\description{
This function computes the bitwise parity of a set of integer, logical, or
raw values, or a list of such vectors.  It is a convenient way to collapse a
set of \code{\linkS4class{UUID}}s into a summary value, regardless of the
order of the \code{UUID}s.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
