% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryString.R
\name{buildQueryString}
\alias{buildQueryString}
\title{Create a URL query string from a list of parameter names and values}
\usage{
buildQueryString(parameters = list())
}
\arguments{
\item{parameters}{A named list with one item (value) for each parameter;
the names of the list items correspond to the names of the parameters}
}
\value{
A character string of the form \code{"param1=value1&param2=value2&..."}
}
\description{
This function creates a URL query string from a list of parameter names and
values.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
