% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GeneHive coercion methods}
\alias{GeneHive coercion methods}
\title{Coerce GeneHive object classes}
\description{
These methods coerce GeneHive S4 objects to objects of other classes, and
vice versa.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
