% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Coercion methods for 'io'}
\alias{Coercion methods for 'io'}
\alias{as}
\alias{coerce,gene.sets,hiveEntityList-method}
\alias{coerce,hiveEntityList,gene.sets-method}
\title{Coerce between GeneHive and io object classes}
\description{
These methods coerce GeneHive S4 objects to objects of classes defined in
package \pkg{io}, or vice versa.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
