% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Coerce S4 to list}
\alias{Coerce S4 to list}
\alias{coerce,S4,list-method}
\title{Coerce an S4 object to a list}
\description{
This method is a convenience function used to coerce an S4 object to a list.
}
\details{
\itemize{
  \item{
    This function will never be called directly, because
    \code{\link[=S4-class]{S4}} is a virtual class;
    however, it is useful to indicate that this function is intended to be
    used generally with S4 objects.
  }
  \item{
    \code{\link{setIs}(\linkS4class{hiveEntity}, S4)} cannot be used
    to ensure that this method is applied to Entity objects, as a conflict
    results for any Entity subclass with a .Data part; as a workaround, the
    definitions for classes that need to use this method are followed by a
    call to \code{setIs} with the parameter \code{class2="S4"} to ensure that
    they inherit the virtual class \code{S4}.
  }
}
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
