% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coerce SimpleList to data.frame}
\alias{coerce SimpleList to data.frame}
\alias{coerce,SimpleList,data.frame-method}
\title{Coerce a SimpleList object to a data frame}
\description{
This method is a convenience function used to coerce a
\code{\linkS4class{SimpleList}} object to a data frame.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
