% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entities.R
\name{Entities}
\alias{Entities}
\alias{addEntity}
\alias{deleteEntity}
\alias{getEntity}
\alias{updateEntity}
\alias{listEntities}
\title{Add, delete, retrieve, update, or list Entity record(s)}
\usage{
addEntity(
  .class,
  ...,
  .permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

deleteEntity(
  .entity_id,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

getEntity(.entity_id, con = hiveConnection())

updateEntity(
  .entity_id,
  ...,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

listEntities(.class, ..., con = hiveConnection())
}
\arguments{
\item{.class}{A character string specifying the class of the \code{Entity} record}

\item{\dots}{Additional arguments specifying fields of the \code{Entity}
record to be added or updated, or fields on which to limit a listing}

\item{.permissions}{A \code{\linkS4class{hivePermissions}} object specifying the permissions to
be used when creating the record}

\item{con}{A \code{\linkS4class{hiveConnection}} object;
if not provided, a new connection will be established}

\item{verbose}{A logical value specifying whether messages should be printed}

\item{.entity_id}{A character string or \code{UUID} specifying the unique identifier of an
\code{Entity} record to be deleted or retrieved}
}
\value{
\describe{
  \item{\code{addEntity}, \code{updateEntity}}{
    If the operation is successful, a \code{\linkS4class{hiveEntity}} object
    (invisibly).
  }
  \item{\code{getEntity}}{
    If the operation is successful, a \code{\linkS4class{hiveEntity}} object.
  }
  \item{\code{deleteEntity}}{
    A logical value stating whether the operation was successful.
  }
  \item{\code{listEntities}}{
    A \code{\linkS4class{hiveEntityList}} object.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add, delete, retrieve or update an \code{Entity}
record, or to list \code{Entity} records.
}
\details{
In functions \code{deleteEntity}, \code{getEntity}, and \code{updateEntity},
the \code{.class} argument does not need to be specified, but if present,
these functions will check that it matches the class of the \code{Entity}
record corresponding to the identifier in argument \code{.entity_id}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
