% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityClasses.R
\name{EntityClasses}
\alias{EntityClasses}
\alias{addEntityClass}
\alias{deleteEntityClass}
\alias{getEntityClass}
\alias{updateEntityClass}
\alias{listEntityClasses}
\title{Add, delete, retrieve, update, or list EntityClass record(s)}
\usage{
addEntityClass(
  filename,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

deleteEntityClass(
  name,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

getEntityClass(name, con = hiveConnection())

updateEntityClass(
  filename,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

listEntityClasses(con = hiveConnection())
}
\arguments{
\item{filename}{A character string specifying the name of a file containing an
\code{EntityClass} definition in JSON}

\item{con}{A \code{\linkS4class{hiveConnection}} object;
if not provided, a new connection will be established}

\item{verbose}{A logical value specifying whether messages should be printed}

\item{name}{A character string specifying the name of an \code{EntityClass}
record to be deleted or retrieved}
}
\value{
\describe{
  \item{\code{addEntityClass}, \code{updateEntityClass}}{
    If the operation is successful, a \code{\linkS4class{hiveEntityClass}}
    object is invisibly returned.
  }
  \item{\code{getEntityClass}}{
    If the operation is successful, a \code{\linkS4class{hiveEntityClass}}
    object is returned.
  }
  \item{\code{deleteEntityClass}}{
    A logical value is returned stating whether the operation was successful.
  }
  \item{\code{listEntityClasses}}{
    A \code{\linkS4class{hiveEntityClassList}} object is returned.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add, delete, retrieve or update an
\code{EntityClass} record, or to list \code{EntityClass} records.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
