% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSets.R
\name{exportFeatureSets}
\alias{exportFeatureSets}
\title{Write FeatureSet(s) to a file}
\usage{
exportFeatureSets(object, filename)
}
\arguments{
\item{object}{A \code{\linkS4class{hiveFeatureSetEntity}} object or a
\code{\linkS4class{hiveEntityList}} object containing one or more
\code{hiveFeatureSetEntity} objects}

\item{filename}{A character string specifying the full path to an output file}
}
\value{
If the operation is successful, the function writes the FeatureSet(s)
to the file specified by \code{filename}; otherwise, the function terminates
with an error message.
}
\description{
This function attempts to write one or more
\code{\linkS4class{hiveFeatureSetEntity}} objects to a file.
}
\seealso{
This function is a wrapper for the \code{\link[io]{qwrite}} function.
Its converse is \code{\link{importFeatureSets}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
