% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedFeatureSets.R
\name{exportWeightedFeatureSet}
\alias{exportWeightedFeatureSet}
\title{Write WeightedFeatureSet to a file}
\usage{
exportWeightedFeatureSet(
  object,
  filename,
  split = TRUE,
  suffixes = c(positive = "_positive", negative = "_negative"),
  verbose = getOption("GeneHive.verbose")
)
}
\arguments{
\item{object}{A \code{\linkS4class{hiveWeightedFeatureSetEntity}} object}

\item{filename}{Character string specifying the full path to an output file}

\item{split}{Logical value specifying whether a \code{WeightedFeatureSet} with weights of
only \code{+1} and \code{-1} should be split into two separate
\code{\linkS4class{hiveFeatureSetEntity}} records when exporting to the file}

\item{suffixes}{Character vector of length 2, with names \code{"positive"} and
\code{"negative"}, indicating the suffixes to be added to the labels of each
\code{FeatureSet} if \code{split=TRUE}}

\item{verbose}{Logical value specifying whether messages should be printed}
}
\value{
If the operation is successful, the function writes the
\code{WeightedFeatureSet} to the file specified by \code{filename};
otherwise, the function terminates with an error message.
}
\description{
This function attempts to write a
\code{\linkS4class{hiveWeightedFeatureSetEntity}}
object to a file.
}
\seealso{
This function is a wrapper for the \code{\link[io]{qwrite}} function.
Its converse is \code{\link{importWeightedFeatureSet}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
