% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{Groups}
\alias{Groups}
\alias{addGroup}
\alias{deleteGroup}
\alias{listGroups}
\title{Add, delete, or list Group record(s)}
\usage{
addGroup(name, con = hiveConnection(), verbose = getOption("GeneHive.verbose"))

deleteGroup(
  name,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

listGroups(con = hiveConnection())
}
\arguments{
\item{name}{A character string specifying the name of a \code{Group}
record to be added or deleted}

\item{con}{A \code{\linkS4class{hiveConnection}} object;
if not provided, a new connection will be established}

\item{verbose}{A logical value specifying whether messages should be printed}
}
\value{
\describe{
  \item{\code{addGroup}}{
    If the operation is successful, a \code{\linkS4class{hiveGroup}}
    object is invisibly returned.
  }
  \item{\code{deleteGroup}}{
    A logical value is returned stating whether the operation was successful.
  }
  \item{\code{listGroups}}{
    A character vector containing all \code{Group} names is returned.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add or delete a \code{Group} record
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
