% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiveBase.R
\name{Hive base functions}
\alias{Hive base functions}
\alias{hiveAdd}
\alias{hiveDelete}
\alias{hiveGet}
\alias{hiveUpdate}
\alias{hiveList}
\title{Add, delete, retrieve, update, or list record(s)}
\usage{
hiveAdd(
  con = hiveConnection(),
  type = c("Entity", "Group", "User"),
  fields,
  verbose = getOption("GeneHive.verbose")
)

hiveDelete(
  con = hiveConnection(),
  type = c("Entity", "EntityClass", "Group"),
  id,
  verbose = getOption("GeneHive.verbose")
)

hiveGet(
  con = hiveConnection(),
  type = c("Entity", "EntityClass", "User", "WorkFileProperties"),
  id
)

hiveUpdate(
  con,
  type = c("Entity", "User", "WorkFileProperties"),
  fields,
  append = TRUE,
  verbose = getOption("GeneHive.verbose")
)

hiveList(
  con,
  type = c("Entity", "EntityClass", "Group", "User", "WorkFileProperties"),
  fields = list(),
  simplify = TRUE
)
}
\arguments{
\item{con}{A \code{\linkS4class{hiveConnection}} object}

\item{type}{A character string specifying the type of the record;
defaults to \code{'Entity'}}

\item{fields}{A list of arguments specifying the unique identifier field of the record to
be updated (see \code{\link{hiveSlotName}}),
and/or any other fields to add or update or on which to limit a listing.
Note:
\itemize{
  \item{
    When limiting a listing, all of the elements of \code{fields} must be of
    length 1.
  }
  \item{
    When adding or listing Entity records, \code{fields} must always include
    a \code{.class} element.
  }
}}

\item{verbose}{A logical value specifying whether messages should be printed}

\item{id}{A vector of length 1 specifying the unique identifier field of the record to
be deleted or retrieved}

\item{append}{A logical value specifying whether to append new elements to array slots
(rather than replace them); defaults to \code{TRUE}}

\item{simplify}{A logical value specifying whether to make listing output more
human-readable (i.e., convert to data frame, show only the lengths of
arrays, convert UUIDs to plain-text descriptions)}
}
\value{
\describe{
  \item{\code{hiveAdd}, \code{hiveUpdate}}{
    If the operation is successful, an S4 object (invisibly).
  }
  \item{\code{hiveGet}}{
    If the operation is successful, an S4 object.
  }
  \item{\code{hiveDelete}}{
    A logical value stating whether the operation was successful (invisibly).
  }
  \item{\code{hiveList}}{
    \describe{
      \item{If \code{simplify} = \code{TRUE}}{
        A data frame containing one row per record and one column per field.
      }
      \item{If \code{simplify} = \code{FALSE}}{
        A \code{\linkS4class{SimpleList}} object
        containing one S4 object per record.
      }
    }
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add, retrieve, or update a GeneHive record, to
delete a Group, EntityClass or Entity record, or to list GeneHive records of
a given class.  In general, they should not be called directly by the user;
the appropriate wrapper functions should be used instead.
}
\details{
\describe{
  \item{\code{hiveAdd}}{
    \itemize{
      \item{
        If a record already exists for the given ID, a warning is produced.
        If the arguments provided differ from those associated with the
        existing record, the function terminates with an error; otherwise,
        the existing record is returned invisibly as an S4 object.
      }
    }
  }
  \item{\code{hiveUpdate}}{
    \itemize{
      \item{
        If a record is unchanged after the update, a warning is produced,
        and the existing record is returned invisibly as an S4 object.
      }
    }
  }
}
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
