% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveConnection-class}
\alias{hiveConnection-class}
\alias{hiveConnection}
\title{Class to contain a connection handle for API calls}
\usage{
hiveConnection(
  username = getOption("GeneHive.username"),
  netrc.file = getOption("GeneHive.netrc.file"),
  proxy = getOption("GeneHive.proxy"),
  verbose = FALSE
)
}
\arguments{
\item{username}{A character string denoting the user name to use when connecting}

\item{netrc.file}{A character string specifying the location of the .netrc file}

\item{proxy}{A logical value specifying whether or not a proxy server is to be used}

\item{verbose}{A logical value specifying whether or not calls made using the
\code{\linkS4class{CURLHandle}} should be verbose}
}
\value{
A \code{\linkS4class{hiveConnection}} object containing the specified
parameters.
}
\description{
This S4 class is a container for a \code{cURL} handle that allows connection
to a GeneHive instance.

This function creates a new \code{\linkS4class{hiveConnection}} object to
be used for making API calls.
}
\section{Extends}{

Directly extends class \code{\linkS4class{CURLHandle}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}

Adam C. Gower \email{agower@bu.edu}
}
