% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveDate-class}
\alias{hiveDate-class}
\alias{hiveDate}
\title{Class to contain a time/date stamp}
\description{
This S4 class is a container for a time/date stamp in a GeneHive record.
}
\section{Extends}{

Directly extends class \code{\link{POSIXct}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
