% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveEntityClass-class}
\alias{hiveEntityClass-class}
\alias{hiveEntityClass}
\title{Class to contain EntityClass records}
\description{
This class is the S4 representation of the EntityClass record type.
It is a container for the definition of an Entity class.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string specifying the name of the record.
Defaults to \code{""}.}

\item{\code{description}}{A character string containing a description of the record.
Defaults to \code{""}.}

\item{\code{variables}}{A \code{\linkS4class{hiveVariableDefinitionCollection}} object containing
any variables associated with the record.}

\item{\code{creator}}{A character string specifying the \code{username} of the User
who created the record. Defaults to \code{""}.}

\item{\code{owner}}{A character string specifying the \code{username} of the User
associated with the record. Defaults to \code{""}.}

\item{\code{group}}{A character string specifying the \code{name} of the Group associated
with the record. Defaults to \code{""}.}

\item{\code{permissions}}{A \code{\linkS4class{hivePermissions}} object specifying the permissions
associated with the record. Defaults to read-only.}

\item{\code{.creation_date}}{A character string specifying the time and date at which the record was
created. Defaults to \code{""}.}

\item{\code{.updated}}{A character string specifying the time and date at which the record was last
updated. Defaults to \code{""}.}
}}

\seealso{
Functions for working with EntityClass records are
described in \code{\link{EntityClasses}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
