% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveEntityClassList-class}
\alias{hiveEntityClassList-class}
\alias{hiveEntityClassList}
\title{Class to contain a list of EntityClass objects}
\usage{
hiveEntityClassList(listData = list())
}
\arguments{
\item{listData}{A list of \code{\linkS4class{hiveEntityClass}} objects}
}
\description{
This class is a container for one or more EntityClass objects.
}
\details{
This constructor function does not require that the \code{listData} argument
be named.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
