% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveEntityList-class}
\alias{hiveEntityList-class}
\alias{hiveEntityList}
\title{Class to contain a list of Entity objects}
\usage{
hiveEntityList(listData = list())
}
\arguments{
\item{listData}{A list of objects of or extending class \code{\linkS4class{hiveEntity}}}
}
\description{
This class is a container for one or more Entity objects.
}
\details{
This constructor function does not require that the \code{listData} argument
be named.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
