% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveFeatureSetEntity-class}
\alias{hiveFeatureSetEntity-class}
\alias{hiveFeatureSetEntity}
\title{Class to describe a set of features}
\description{
This class is the S4 representation of the Entity class FeatureSet.
It is a container for an unweighted set of features (e.g., a gene set).
}
\section{Slots}{

\describe{
\item{\code{.entity_id}}{A \code{\linkS4class{UUID}} automatically computed from the features
when adding a new FeatureSet record.}
}}

\section{Extends}{

Directly extends classes \code{\linkS4class{hiveEntity}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
