% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveFeatureSpaceEntity-class}
\alias{hiveFeatureSpaceEntity-class}
\alias{hiveFeatureSpaceEntity}
\title{Class to describe a feature space}
\description{
This class is the S4 representation of the Entity class FeatureSpace.
It is a container for \emph{either} a set of features or a regular expression
defining a feature space (e.g., Entrez Gene IDs, Affymetrix probeset IDs for
a specific microarray platform, etc.).
}
\section{Slots}{

\describe{
\item{\code{.entity_id}}{A \code{\linkS4class{UUID}} automatically computed from the contents of the
record.}
}}

\section{Extends}{

Directly extends class \code{\linkS4class{hiveEntity}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
