% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hivePermissions-class}
\alias{hivePermissions-class}
\alias{hivePermissions}
\title{Class to contain permissions for a record}
\usage{
hivePermissions(...)
}
\arguments{
\item{\dots}{Optional arguments containing permission information to be used to populate
slots of output}
}
\description{
This S4 class is a container for a set of permissions associated with a
GeneHive record.
}
\section{Slots}{

\describe{
\item{\code{group}}{A character vector specifying the permissions that are granted
to members of the Group associated with the record:
may include \code{"R"}ead, \code{"U"}pdate, or \code{"D"}elete.
Defaults to an empty vector (i.e., no permissions).}

\item{\code{other}}{A character vector specifying the permissions that are granted
to all Users:
may include \code{"R"}ead, \code{"U"}pdate, or \code{"D"}elete.
Defaults to an empty vector (i.e., no permissions).}
}}

\author{
Adam C. Gower \email{agower@bu.edu}
}
