% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiveSlots.R
\name{Get or set hive S4 object slots}
\alias{Get or set hive S4 object slots}
\alias{objectCreationDate}
\alias{objectCreator}
\alias{objectDescription}
\alias{objectGroup}
\alias{objectId}
\alias{objectLastUpdated}
\alias{objectName}
\alias{objectOwner}
\alias{objectPermissions}
\alias{objectDescription<-}
\alias{objectGroup<-}
\alias{objectId<-}
\alias{objectName<-}
\alias{objectOwner<-}
\alias{objectPermissions<-}
\title{Get or set various slots of a hive S4 objects}
\usage{
objectCreationDate(object)

objectCreator(object)

objectDescription(object)

objectGroup(object)

objectId(object)

objectLastUpdated(object)

objectName(object)

objectOwner(object)

objectPermissions(object)

objectDescription(object) <- value

objectGroup(object) <- value

objectId(object) <- value

objectName(object) <- value

objectOwner(object) <- value

objectPermissions(object) <- value
}
\arguments{
\item{object}{An R object}

\item{value}{The value that the slot should take}
}
\value{
\describe{
  \item{
    objectCreationDate, objectCreator, objectDescription, objectGroup,
    objectName, objectOwner
  }{
    A character vector of length 1.
  }
  \item{objectId}{
    The unique identifier is returned; for a
    \code{\linkS4class{hiveEntity}}, this will be a
    \code{\linkS4class{UUID}} object; for a
    \code{\linkS4class{hiveWorkFileProperties}} object, this will be a
    \code{\linkS4class{hiveWorkFileID}} object; and for
    \code{\linkS4class{hiveEntityClass}}, \code{\linkS4class{hiveGroup}} or
    \code{\linkS4class{hiveUser}} objects, this will be a character string.
  }
  \item{objectPermissions}{
    A \code{linkS4class{hivePermissions}} object.
  }
}
}
\description{
These are convenience functions for getting and setting various slots of
an S4 representation of a hive record.
}
\seealso{
The \code{\link{hiveSlotName}} function returns the name of the slot that
corresponds to these fields, rather than the field itself.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
