% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveUser-class}
\alias{hiveUser-class}
\alias{hiveUser}
\title{Class to contain data about a User}
\usage{
hiveUser(...)
}
\arguments{
\item{\dots}{Arguments containing user information to be used to populate slots of output}
}
\description{
This class is the S4 representation of the User record type.
It is a container for data about a given user.
}
\section{Slots}{

\describe{
\item{\code{username}}{A character string specifying the name of the user.
Defaults to \code{""}.}

\item{\code{password}}{A character string specifying the password of the user.
Defaults to \code{NA}.}

\item{\code{group}}{A character string specifying the name of the default Group associated with
the user. Defaults to \code{""}.}

\item{\code{groups}}{A character array specifying the name(s) of the Group(s) associated with
the user. Defaults to \code{""}.}

\item{\code{superuser}}{A logical value specifying whether the user is a superuser.
Defaults to \code{FALSE}.}

\item{\code{confirmed}}{A logical value specifying whether the user has been confirmed.
Defaults to \code{FALSE}.}

\item{\code{email}}{A character string specifying the email address of the user.
Defaults to \code{""}.}

\item{\code{firstName}}{A character string specifying the first name of the user.
Defaults to \code{""}.}

\item{\code{lastName}}{A character string specifying the last name of the user.
Defaults to \code{""}.}

\item{\code{dateJoined}}{A character string specifying the date that the User record was
created. Automatically populated when the record is created.}

\item{\code{lastLogin}}{A character string specifying the timestamp of the User's last
authentication. Automatically updated by the system.}

\item{\code{untrashedFileUsage}}{A numeric value specifying the file usage of the user.
Automatically updated by the system.}

\item{\code{token}}{A character string specifying a token that can be used for password-free
authentication. Automatically populated when the record is created.}

\item{\code{active}}{A logical value specifying whether the user is active.
Defaults to \code{FALSE}.}
}}

\seealso{
\describe{
  \item{group}{
    This slot corresponds to the \code{\linkS4class{hiveGroup}} object slot
    \code{name}.
  }
}
Functions for working with User records are described in \code{\link{Users}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
