% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveUserList-class}
\alias{hiveUserList-class}
\alias{hiveUserList}
\title{Class to contain a list of User objects}
\usage{
hiveUserList(listData = list())
}
\arguments{
\item{listData}{A list of objects of class \code{\linkS4class{hiveUser}}}
}
\description{
This class is a container for one or more User objects.
}
\details{
This constructor function does not require that the \code{listData} argument
be named.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
