% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiveUtilities.R
\name{Hive utility functions}
\alias{Hive utility functions}
\alias{hiveApp}
\alias{hiveExists}
\alias{hiveKeyFields}
\alias{hiveLabelFields}
\alias{hiveS4Class}
\alias{hiveSlotName}
\title{Utility functions used by other functions}
\usage{
hiveApp(
  type = c("Entity", "EntityClass", "Group", "User", "WorkFileProperties")
)

hiveExists(
  id,
  type = c("Entity", "EntityClass", "Group", "User", "WorkFileProperties"),
  con = hiveConnection()
)

hiveKeyFields(class, con = hiveConnection())

hiveLabelFields(class, con = hiveConnection())

hiveS4Class(
  type = c("Entity", "EntityClass", "Group", "User", "WorkFileProperties"),
  class
)

hiveSlotName(
  Class,
  attribute = c("creationDate", "creator", "description", "group", "id", "name",
    "owner", "permissions", "updated")
)
}
\arguments{
\item{type}{A character string representing a record type;
defaults to \code{"Entity"}}

\item{id}{A character string specifying a record identifier}

\item{con}{A \code{\linkS4class{hiveConnection}} object}

\item{class}{A character string representing an Entity class;
in function \code{hiveS4Class}, this is an optional argument
that is used to determine S4 class if type = \code{"Entity"}}

\item{Class}{A character string representing an S4 object class}

\item{attribute}{A character string representing an attribute (e.g., "id", "name") for which
the corresponding slot name is desired}
}
\value{
\describe{
  \item{hiveApp}{
    A character vector of length 1 specifying the app name to be used in the
    API URL
  }
  \item{hiveExists}{
    A logical value specifying whether a record of the given type exists
    with the given identifier
  }
  \item{hiveKeyFields}{
    A character vector containing the names of any \code{'key'} fields of an
    \code{Entity} class (i.e., those used to compute the UUID of the
    record), or, if no key fields exist for the class, a character vector of
    length 0
  }
  \item{hiveS4Class}{
    If the input corresponds to a valid S4 object class, it is returned;
    otherwise, \code{NULL}.
  }
  \item{hiveSlotName}{
    A character vector of length 1 containing the name of the slot that
    corresponds to the specified attribute
  }
}
}
\description{
These convenience functions perform low-level tasks used by other GeneHive
functions.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
