% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveVariableDefinition-class}
\alias{hiveVariableDefinition-class}
\alias{hiveVariableDefinition}
\title{Class to contain VariableDefinition records}
\description{
This class is the S4 representation of the VariableDefinition
record type.  It is a container for a single variable associated with an
EntityClass record.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string specifying the name of the record.
Defaults to \code{""}.}

\item{\code{category}}{A character string describing the category of the record.
Defaults to \code{NA}.}

\item{\code{type}}{A one-character string specifying the type of the record.
Defaults to \code{""}.  May take the value \code{'C'} (Code),
\code{'D'} (Date), \code{'E'} (Entity), \code{'F'} (Float),
\code{'I'} (Integer), \code{'S'} (String), \code{'T'} (Text),
or \code{'W'} (WorkFile).}

\item{\code{codes}}{A character vector specifying the allowed values that the variable may take
(if \code{type} is \code{'C'}).
Defaults to an empty vector.}

\item{\code{entity_class_name}}{A character string specifying the name of the EntityClass that the
variable must belong to (if \code{type} is \code{'E'}).
Defaults to \code{NA}.}

\item{\code{is_array}}{A logical value specifying whether the variable is an array.
Defaults to \code{FALSE}.}
}}

\seealso{
VariableDefinition objects are contained in
\code{\linkS4class{hiveVariableDefinitionCollection}} objects.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
