% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveVariableDefinitionCollection-class}
\alias{hiveVariableDefinitionCollection-class}
\alias{hiveVariableDefinitionCollection}
\title{Class to contain VariableDefinitionCollection records}
\usage{
hiveVariableDefinitionCollection(listData = list())
}
\arguments{
\item{listData}{A list of hiveVariableDefinition objects}
}
\description{
This class is the S4 representation of the VariableDefinitionCollection
record type.
It is a container for the variables and metadata associated with a collection
of variable definitions in an EntityClass record.
}
\section{Slots}{

\describe{
\item{\code{elementType}}{Contains the character string \code{"VariableDefinition"}.}

\item{\code{listData}}{A list of \code{\linkS4class{hiveVariableDefinition}} objects.}
}}

\section{Extends}{

Directly extends class \code{\linkS4class{SimpleList}}.
}

\seealso{
VariableDefinitionCollection objects are contained in
\code{\linkS4class{hiveEntityClass}} objects.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
