% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveWeightedFeatureSetEntity-class}
\alias{hiveWeightedFeatureSetEntity-class}
\alias{hiveWeightedFeatureSetEntity}
\title{Class to describe a set of features and weights}
\description{
This class is the S4 representation of the Entity class
WeightedFeatureSet.  It is a container for a set of features and their
associated weights (e.g., a ranked list or a biomarker).
}
\section{Slots}{

\describe{
\item{\code{.entity_id}}{A \code{\linkS4class{UUID}} automatically computed from the features and
weights when adding a new WeightedFeatureSet record.}
}}

\section{Extends}{

Directly extends class \code{\linkS4class{hiveEntity}}.
}

\seealso{
The \code{\linkS4class{hiveFeatureSetEntity}} class is used to contain a set
of unweighted features.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
