% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveWorkFileID-class}
\alias{hiveWorkFileID-class}
\alias{hiveWorkFileID}
\title{Class to contain a WorkFile identifier}
\description{
This class is a container for the unique identifier of a WorkFile.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A character string specifying the unique identifier of the WorkFile
(most likely a character representation of an integer, though this is not
guaranteed).}
}}

\author{
Adam C. Gower \email{agower@bu.edu}
}
