% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveWorkFileIDList-class}
\alias{hiveWorkFileIDList-class}
\alias{hiveWorkFileIDList}
\title{Class to contain a list of WorkFile identifiers}
\usage{
hiveWorkFileIDList(listData = list())
}
\arguments{
\item{listData}{A list of \code{\linkS4class{hiveWorkFileID}} objects}
}
\description{
This class is a container for the unique identifier of one or more
WorkFiles.
}
\details{
This constructor function does not require that the \code{listData} argument
be named.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
