% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSets.R
\name{importFeatureSets}
\alias{importFeatureSets}
\title{Upload FeatureSet(s) from a file}
\usage{
importFeatureSets(
  filename,
  featureSpace,
  ...,
  na.rm = TRUE,
  .permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)
}
\arguments{
\item{filename}{A character string specifying the full path to a file containing one or more
FeatureSets}

\item{featureSpace}{An optional character string or \code{\linkS4class{UUID}} specifying the
\code{\linkS4class{hiveFeatureSpaceEntity}} associated with the FeatureSet.
If present, the FeatureSet will be checked against it for validity;
if missing, an attempt is made to impute the \code{FeatureSpace}
automatically.}

\item{\dots}{Optional arguments to \code{\link{addEntity}}, e.g., \code{description}
or \code{species}}

\item{na.rm}{A logical value specifying how to deal with features with NA values:
if \code{TRUE}, such features are removed from input;
if \code{FALSE}, the function will terminate with an error message}

\item{.permissions}{A \code{\linkS4class{hivePermissions}} object specifying the permissions
to be used when creating the record}

\item{con}{A \code{\linkS4class{hiveConnection}} object;
if not provided, a new connection will be established}

\item{verbose}{A logical value specifying whether messages should be printed}
}
\value{
If the operation is successful, the function invisibly returns a
\code{\linkS4class{hiveEntityList}} object containing one or more
FeatureSet objects;
otherwise, the function terminates with an error message.
}
\description{
This function attempts to upload one or more
\code{\linkS4class{hiveFeatureSetEntity}} records from a file to a GeneHive.
}
\seealso{
This function is a wrapper for the \code{\link[io]{qread}} function.
Its converse is \code{\link{exportFeatureSets}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
