% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPublication.R
\name{importPublication}
\alias{importPublication}
\title{Create a Publication record using a PubMed identifier}
\usage{
importPublication(
  PMID,
  .permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)
}
\arguments{
\item{PMID}{A numeric or character vector of length one specifying a PubMed identifier}

\item{.permissions}{A \code{\linkS4class{hivePermissions}} object specifying the permissions to
be used when creating or updating the record}

\item{con}{A \code{\linkS4class{hiveConnection}} object;
if not provided, a new connection will be established}

\item{verbose}{A logical value specifying whether messages should be printed}
}
\value{
If the operation is successful, a \code{\linkS4class{hiveEntity}}
object corresponding to the Publication record is invisibly returned;
otherwise, the function terminates with an error message.
}
\description{
This function adds or updates a single Publication record using metadata
associated with a PubMed identifier.
}
\details{
The fields of the Publication record are populated from the PubMed XML
record. If a DOI is present, it is added; otherwise, that field is left
blank.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
