% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedFeatureSets.R
\name{importWeightedFeatureSet}
\alias{importWeightedFeatureSet}
\title{Upload WeightedFeatureSet from a file}
\usage{
importWeightedFeatureSet(
  filename,
  featureSpace,
  ...,
  na.rm = TRUE,
  .permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)
}
\arguments{
\item{filename}{Character string specifying the full path to a file containing a
\code{WeightedFeatureSet}}

\item{featureSpace}{An optional \code{\linkS4class{UUID}} specifying the
\code{\linkS4class{hiveFeatureSpaceEntity}} associated with the
\code{WeightedFeatureSet}. If present, the \code{WeightedFeatureSet} will be
checked against it for validity; if missing, an attempt is made to impute
the \code{FeatureSpace} automatically.}

\item{\dots}{Optional arguments to \code{\link{addEntity}},
e.g., \code{description}}

\item{na.rm}{Logical value specifying how to deal with features with NA weights and/or
names: if \code{TRUE}, such features are removed from input; if \code{FALSE},
the function terminates with an error message}

\item{.permissions}{A \code{\linkS4class{hivePermissions}} object specifying the permissions to
be used when creating the record}

\item{con}{A \code{\linkS4class{hiveConnection}} object;}

\item{verbose}{Logical value specifying whether messages should be printed}
}
\value{
If the operation is successful, the function invisibly returns a
\code{\linkS4class{hiveWeightedFeatureSetEntity}} object;
otherwise, the function
terminates with an error message.
}
\description{
This function attempts to upload a
\code{\linkS4class{hiveWeightedFeatureSetEntity}} record from a file.
}
\seealso{
This function is a wrapper for the \code{\link[io]{qread}} function.
Its converse is \code{\link{exportWeightedFeatureSet}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
