% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{imputeFeatureSpace}
\alias{imputeFeatureSpace}
\alias{imputeFeatureSpace,hiveEntityList-method}
\alias{imputeFeatureSpace,hiveFeatureSetEntity-method}
\alias{imputeFeatureSpace,hiveWeightedFeatureSetEntity-method}
\alias{imputeFeatureSpace,character-method}
\title{Impute the FeatureSpace associated with an object}
\usage{
imputeFeatureSpace(object, con = hiveConnection())

\S4method{imputeFeatureSpace}{hiveEntityList}(object, con = hiveConnection())

\S4method{imputeFeatureSpace}{hiveFeatureSetEntity}(object, con = hiveConnection())

\S4method{imputeFeatureSpace}{hiveWeightedFeatureSetEntity}(object, con = hiveConnection())

\S4method{imputeFeatureSpace}{character}(object, con = hiveConnection())
}
\arguments{
\item{object}{Either a \code{\linkS4class{hiveFeatureSetEntity}} or
\code{\linkS4class{hiveWeightedFeatureSetEntity}} object, a
\code{\linkS4class{hiveEntityList}} containing one or more of these objects,
or a \code{character} vector}

\item{con}{A \code{\linkS4class{hiveConnection}} object}
}
\value{
If successful, each class-specific method returns a
\code{\linkS4class{UUIDList}} containing the following output for each
object in the \code{hiveEntityList}:
\itemize{
  \item{
    If the features in a given record do not match any
    \code{FeatureSpace}s, a nil \code{\linkS4class{UUID}} is returned,
    with a warning.
  }
  \item{
    If the features match exactly one \code{FeatureSpace},
    the UUID of that \code{FeatureSpace} is returned.
  }
  \item{
    If the features match more than one \code{FeatureSpace},
    the \code{UUID} of the smallest \code{FeatureSpace} is returned.
  }
}
Otherwise, the function terminates with an error message.
}
\description{
This is a generic function for imputing the \code{FeatureSpace} associated
with one or more sets of features.
}
\details{
If a given set of features matches one or more FeatureSpaces consisting of a
set of features, as well as one or more FeatureSpaces consisting of regular
expressions, precedence is given to the FeatureSpace(s) that consist of a
set of features.
}
\section{Methods}{

Class-specific methods exist for:
\itemize{
  \item{\code{\linkS4class{hiveFeatureSetEntity}}}
  \item{\code{\linkS4class{hiveWeightedFeatureSetEntity}}}
  \item{\code{\linkS4class{hiveEntityList}}}
  \item{\code{\linkS4class{character}}}
}
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
