% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Initialize a hiveConnection}
\alias{Initialize a hiveConnection}
\alias{initialize,hiveConnection-method}
\title{Initialize a hiveConnection object}
\usage{
\S4method{initialize}{hiveConnection}(.Object, ...)
}
\arguments{
\item{.Object}{An uninitialized \code{\linkS4class{hiveConnection}} object}

\item{\dots}{Parameters used to populate the slots of \code{.Object}}
}
\value{
An initialized \code{\linkS4class{hiveConnection}} object.
}
\description{
This method is required to initialize the external pointer \code{ref} in a
\code{\linkS4class{hiveConnection}} object
(i.e., \code{\linkS4class{CURLHandle}}).
Note: in package \code{RCurl}, this is accomplished by
\code{\link{getCurlHandle}}, not by \code{\link{new}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
