% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Initialize a hiveEntity}
\alias{Initialize a hiveEntity}
\alias{initialize,hiveEntity-method}
\title{Initialize a hiveEntity object}
\usage{
\S4method{initialize}{hiveEntity}(.Object, ...)
}
\arguments{
\item{.Object}{An uninitialized \code{\linkS4class{hiveEntity}} object}

\item{\dots}{Parameters used to populate the slots of \code{.Object}}
}
\value{
\itemize{
  \item{
    If any non-array variables in \code{.Object} are of length > 1,
    the function terminates with an error.
  }
  \item{
    Otherwise, an initialized \code{\linkS4class{hiveEntity}} object
    is returned.
  }
}
}
\description{
The \code{hiveEntity}-specific method performs the following tasks:
\itemize{
  \item{
    Initializes \code{.Object} using the parameters defined in the \dots
    argument
  }
  \item{
    Computes the \code{.entity_id} slot from the key fields (if any)
  }
  \item{
    If the key fields are empty,
    the slot \code{.entity_id} remains a nil UUID (default)
  }
}
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
