% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isJSON.R
\name{isJSON}
\alias{isJSON}
\title{Test whether a string encodes valid JSON}
\usage{
isJSON(x)
}
\arguments{
\item{x}{A character string}
}
\value{
A logical value specifying whether the stirng contains valid JSON.
}
\description{
This function tests whether a given string contains syntactically correct
JSON.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
