% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length}
\alias{length}
\alias{length,hiveFeatureSpaceEntity-method}
\alias{length,hiveFeatureSetEntity-method}
\alias{length,hiveWeightedFeatureSetEntity-method}
\title{Get the length of a hiveEntity}
\usage{
\S4method{length}{hiveFeatureSpaceEntity}(x)

\S4method{length}{hiveFeatureSetEntity}(x)

\S4method{length}{hiveWeightedFeatureSetEntity}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{hiveEntity}} object}
}
\value{
An integer corresponding to the length of the internal vector.
}
\description{
These methods get the length of a \code{\linkS4class{hiveEntity}} containing
a vector of identifiers or features.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
