% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToHiveS4.R
\name{List to hive S4 object}
\alias{List to hive S4 object}
\alias{listToHiveS4}
\title{Convert a list of fields to a hive S4 object}
\usage{
listToHiveS4(Class, x)
}
\arguments{
\item{Class}{A character string representing an S4 object class}

\item{x}{A list containing fields that will be coerced to slots of the S4 object}
}
\value{
An object of the class specified in \code{Class}.
}
\description{
This function creates an S4 object from a list of fields that correspond to
slots of a hive S4 object, performing a recursive validity check.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
