% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passwords.R
\name{Passwords}
\alias{Passwords}
\alias{changePassword}
\alias{checkPassword}
\alias{randomPassword}
\alias{resetPassword}
\title{Change, check, or reset passwords}
\usage{
changePassword(
  username = getOption("GeneHive.username"),
  random = TRUE,
  netrc.file = getOption("GeneHive.netrc.file"),
  verbose = getOption("GeneHive.verbose")
)

checkPassword(username = getOption("GeneHive.username"), password)

randomPassword(nchars = 20)

resetPassword(username, prefix, verbose = getOption("GeneHive.verbose"))
}
\arguments{
\item{username}{A character string containing the name of the User;
defaults to \code{getOption("GeneHive.username")}}

\item{random}{A logical value specifying whether a random password should be generated;
defaults to \code{TRUE}. If available, \code{/dev/urandom} is used to
generate random passwords; otherwise, \code{\link{sample}} is used.}

\item{netrc.file}{A character string specifying the location of the .netrc file}

\item{verbose}{A logical value specifying whether messages should be printed}

\item{password}{A character string containing the password that will be checked for a User}

\item{nchars}{The length of the password to be generated, in characters}

\item{prefix}{An optional character string containing the URL prefix of a landing page to
use in email template. This will have \code{"?user={user}&token={token}"}
appended in the actual template.
If this is absent, then the template for direct API use will be sent instead
of the template linking to a landing page.}
}
\value{
\describe{
  \item{\code{changePassword}}{
    \code{TRUE} (invisibly) if the password was successfully
    changed; otherwise, the function terminates with an error.
  }
  \item{\code{checkPassword}}{
    \code{TRUE} if the username and password are valid,
    \code{FALSE} otherwise.
  }
  \item{\code{randomPassword}}{
    The password (invisibly).
  }
  \item{\code{resetPassword}}{
    The HTTP response body of the request (invisibly).
  }
}
}
\description{
These functions allow Users to change their password or (for superusers)
other Users' passwords, to check a username/password combination, to
generate a random password, or (for superusers) to send a password reset
email to a User.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
