% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityS4Class.R
\name{refreshEntityS4Class}
\alias{refreshEntityS4Class}
\title{Refresh Entity S4 class definition}
\usage{
refreshEntityS4Class(class, verbose = getOption("GeneHive.verbose"))
}
\arguments{
\item{class}{A character string specifying the Entity class}

\item{verbose}{Logical value specifying whether messages should be printed}
}
\value{
\itemize{
  \item{
    If the S4 class definition is successfully set or refreshed, the
    class constructor function (invisibly).
  }
  \item{
    If the S4 class definition exists in two or more namespaces (not
    including \code{.GlobalEnv}) the function terminates with an error.
  }
  \item{
    If a single S4 class definition exists, but is sealed,
    the function terminates with an error.
  }
}
}
\description{
This function refreshes the local S4 class definition corresponding to a
given Entity class.
}
\details{
\itemize{
  \item{
    If the S4 class definition does not exist, it will be set in
    the \code{\link{.GlobalEnv}} environment.
  }
  \item{
    If the S4 class definition exists in one namespace, is unsealed, and
    does not match the current EntityClass definition, it will be refreshed.
  }
  \item{
    If the S4 class definition exists in two namespaces, one of which is
    \code{.GlobalEnv}, the definition in \code{.GlobalEnv} will be removed.
    If the remaining definition is unsealed and does not match the current
    EntityClass definition, it will be refreshed.
  }
}
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
