% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityS4Class.R
\name{setEntityS4Class}
\alias{setEntityS4Class}
\title{Set an Entity S4 class definition}
\usage{
setEntityS4Class(entityClassDef, where = .GlobalEnv)
}
\arguments{
\item{entityClassDef}{A \code{\linkS4class{hiveEntityClass}} object}

\item{where}{An environment in which to create the S4 class definition if needed}
}
\value{
The function calls \link{setClass} to create the S4 class definition within
its namespace, or, if it does not already exist, to create the S4 class
definition within the environment specified in argument \code{where}.
}
\description{
This function creates a local S4 class definition corresponding to a given
Entity class definition.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
