% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{S4 show methods}
\alias{S4 show methods}
\alias{show,hiveDate-method}
\alias{show,hiveEntity-method}
\alias{show,hiveEntityList-method}
\alias{show,hiveEntityClass-method}
\alias{show,hiveEntityClassList-method}
\alias{show,hivePermissions-method}
\alias{show,hiveUser-method}
\alias{show,hiveWorkFileProperties-method}
\title{Methods for showing S4 objects}
\usage{
\S4method{show}{hiveDate}(object)

\S4method{show}{hiveEntity}(object)

\S4method{show}{hiveEntityList}(object)

\S4method{show}{hiveEntityClass}(object)

\S4method{show}{hiveEntityClassList}(object)

\S4method{show}{hivePermissions}(object)

\S4method{show}{hiveUser}(object)

\S4method{show}{hiveWorkFileProperties}(object)
}
\arguments{
\item{object}{An R object}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
These methods show the contents of various S4 objects in a human-readable
format.
}
\section{Methods}{

Class-specific methods exist for:
\itemize{
  \item \code{\linkS4class{hiveDate}}
  \item \code{\linkS4class{hiveEntity}}
  \item \code{\linkS4class{hiveEntityList}}
  \item \code{\linkS4class{hiveEntityClass}}
  \item \code{\linkS4class{hiveEntityClassList}}
  \item \code{\linkS4class{hivePermissions}}
  \item \code{\linkS4class{hiveUser}}
  \item \code{\linkS4class{hiveWorkFileProperties}}
}
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
