% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopIfHiveError.R
\name{stopIfHiveError}
\alias{stopIfHiveError}
\title{Terminate if an HTTP response is an error message}
\usage{
stopIfHiveError(x)
}
\arguments{
\item{x}{A variable returned by \code{\link{httpRequest}}}
}
\value{
If the argument \code{x} is of class \code{\link{simpleError}}, the function
processes the error message as needed and terminates with the error;
otherwise, the argument \code{x} is returned.
}
\description{
This is a convenience function to terminate if the result of an HTTP request
contains an error message.
}
\details{
The system may return a JSON-encoded list containing a single element,
named \code{error}, containing an error message.  If so, it is extracted to
an error message during termination (see below).
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
