% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{translateFeatures}
\alias{translateFeatures}
\title{Translate a set of features from one feature space to another}
\usage{
translateFeatures(object, translator, ...)
}
\arguments{
\item{object}{An object containing a set of features that will be translated from one
feature space to another}

\item{translator}{An object containing a table, etc., that may be used to translate the
features in \code{object}}

\item{\dots}{Additional arguments to methods of \code{translateFeatures},
e.g., \code{species}}
}
\value{
Each method returns either an S4 object of the same class as \code{object},
or a \code{\linkS4class{SimpleList}} of such objects.
}
\description{
This is a generic function for translating a set of features from one
feature space to another.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
