% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkFileProperties.R
\name{WorkFileProperties}
\alias{WorkFileProperties}
\alias{getWorkFileProperties}
\alias{updateWorkFileProperties}
\alias{listWorkFileProperties}
\alias{listWorkFiles}
\title{Retrieve, update, or list WorkFileProperties record(s)}
\usage{
getWorkFileProperties(id, con = hiveConnection())

updateWorkFileProperties(
  id,
  ...,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

listWorkFileProperties(
  ...,
  isTrashed = FALSE,
  con = hiveConnection(),
  simplify = TRUE
)

listWorkFiles(..., isTrashed = FALSE, con = hiveConnection(), simplify = TRUE)
}
\arguments{
\item{id}{A \code{\linkS4class{hiveWorkFileID}} object specifying the unique identifier
of a \code{WorkFileProperties} record.
Automatically populated when a \code{WorkFile} is uploaded.}

\item{con}{A \code{\linkS4class{hiveConnection}} object;
if not provided, a new connection will be established}

\item{\dots}{Additional arguments specifying fields of the \code{WorkFileProperties}
record to be updated, or fields on which to limit a listing}

\item{verbose}{A logical value specifying whether messages should be printed}

\item{isTrashed}{A logical value specifying how to limit the result with respect to trashed
records.  If \code{FALSE} (default), only untrashed records are returned;
if \code{TRUE}, only trashed records are returned; if \code{NA}, both trashed
and untrashed records are returned.}

\item{simplify}{A logical value specifying whether to return the listing as a data frame}
}
\value{
\describe{
  \item{\code{getWorkFileProperties}}{
    If the operation is successful,
    a \code{\linkS4class{hiveWorkFileProperties}} object.
  }
  \item{\code{updateWorkFileProperties}}{
    If the operation is successful,
    a \code{\linkS4class{hiveWorkFileProperties}} object (invisibly).
  }
  \item{\code{listWorkFileProperties}}{
    \describe{
      \item{If \code{simplify} = \code{TRUE}}{
        A data frame containing one row per record and one column per field.
      }
      \item{If \code{simplify} = \code{FALSE}}{
        A \code{\linkS4class{SimpleList}} object
        containing one \code{\linkS4class{hiveWorkFileProperties}} object
        per record.
      }
    }
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to retrieve a \code{WorkFileProperties} record
or to list \code{WorkFileProperties} records.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
