% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkFiles.R
\name{WorkFiles}
\alias{WorkFiles}
\alias{addWorkFile}
\alias{getWorkFile}
\alias{getWorkFileAsObject}
\alias{importWorkFiles}
\alias{importWorkFileArchive}
\alias{storeObjectAsWorkFile}
\alias{validWorkFileId}
\title{Functions for working with WorkFiles}
\usage{
addWorkFile(
  filename,
  fileType = file_ext(sub("\\\\.gz$", "", filename)),
  compress = TRUE,
  group,
  token,
  permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

getWorkFile(
  id,
  filename,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

getWorkFileAsObject(
  id,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

importWorkFiles(
  filenames,
  ...,
  permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

importWorkFileArchive(
  filename,
  ...,
  permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

storeObjectAsWorkFile(
  x,
  permissions = getOption("GeneHive.permissions"),
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

validWorkFileId(id, con = hiveConnection())
}
\arguments{
\item{filename}{A character string specifying the full path to a file; for
\code{importWorkFileArchive}, this must be a .tar, .tgz, .tar.gz or .zip file}

\item{fileType}{A character string specifying the file type (e.g., extension) of the 
\code{WorkFile}}

\item{compress}{A logical value specifying whether to gzip-compress the \code{WorkFile} on
upload; if the contents of the \code{WorkFile} are already gzipped, this is
ignored}

\item{group}{A character string specifying the name of the \code{Group} associated with
the \code{WorkFile}. If missing, the hive automatically uses the group
belonging to the user uploading the \code{WorkFile}.}

\item{token}{A character string specifying a token that can be used for password-free 
authentication. Automatically created when a \code{WorkFile} is uploaded.}

\item{permissions}{A \code{\linkS4class{hivePermissions}} object specifying the permissions to 
be used when creating the record}

\item{con}{A \code{\linkS4class{hiveConnection}} object; if not provided, a new
connection will be established}

\item{verbose}{A logical value specifying whether messages should be printed}

\item{id}{A \code{\linkS4class{hiveWorkFileID}} object specifying the unique
identifier of a \code{WorkFileProperties} record.
Automatically created when a \code{WorkFile} is uploaded.}

\item{filenames}{A character vector specifying the full path to one or more files}

\item{\dots}{Additional arguments specifying \code{WorkFileProperties} fields on which to 
limit a listing}

\item{x}{An R object to be uploaded as a \code{WorkFile}}
}
\value{
\describe{
  \item{\code{addWorkFile}}{
    If a \code{WorkFile} exists with the same hash as the input file, it is
    assumed that they are identical; the existing record will be returned
    invisibly as a \code{linkS4class{hiveWorkFileProperties}} object, and a
    warning message will be issued. If a \code{WorkFile} does not exist with
    the same MD5 hash as the input file, the file is uploaded and the
    resulting \code{WorkFile} is returned invisibly as a
    \code{linkS4class{hiveWorkFileProperties}} object.
  }
  \item{\code{getWorkFile}}{
    No value is returned.
  }
  \item{\code{importWorkFiles, importWorkFileArchive}}{
    If the operation is successful, a
    \code{\linkS4class{hiveWorkFileIDList}} object is returned invisibly;
    otherwise, the function terminates with an error message.
  }
  \item{\code{storeObjectAsWorkFile}}{
    If the operation is successful, a
    \code{\linkS4class{hiveWorkFileProperties}} object is invisibly returned.
  }
  \item{\code{getWorkFileAsObject}}{
    If the operation is successful, an R object is invisibly returned:
    either the R object encoded in the file (if the WorkFile is an
    \link[=serialize]{RDS} file) or the contents of the WorkFile (as
    a raw vector).
  }
  \item{\code{listWorkFiles}}{
    A data frame containing one row per record and one column per field.
  }
  \item{\code{validWorkFileId}}{
    A logical value specifying whether the string corresponds to a
    WorkFile identifier.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add or retrieve a single \code{WorkFile}, to list
\code{WorkFile} records, to add a set of \code{WorkFile}s (specified via a
set of filenames or in an archive), to pass data between R objects and
\code{WorkFile}s, or to test whether a character string represents a valid
\code{WorkFile} identifier.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
