% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveGroup-class}
\alias{hiveGroup-class}
\alias{hiveGroup}
\title{Class to contain data about a Group}
\description{
This class is the S4 representation of the Group record type.
It is a container for data about a given group of Users.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string specifying the name of the group.
Defaults to \code{""}.}
}}

\seealso{
Functions for working with Group records in the hive are described in
\code{\link{Groups}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
