% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiveURL.R
\name{hiveURL}
\alias{hiveURL}
\title{Construct a hive API URL}
\usage{
hiveURL(
  ...,
  query,
  hostname = getOption("GeneHive.hostname"),
  https = getOption("GeneHive.https"),
  api.base.path = getOption("GeneHive.api.base.path")
)
}
\arguments{
\item{\dots}{Character strings, integers, and/or numeric values denoting items to be used
to construct the path}

\item{query}{An optional named list of strings that specify the query parameters}

\item{hostname}{A character string specifying the hostname (IP address) of the hive; defaults
to the contents of the environment variable \code{HIVE_HOSTNAME} (if set) or
\code{"localhost"} (if not set)}

\item{https}{A logical value specifying whether a secure HTTP connection should be used;
defaults to the contents of the environment variable \code{HIVE_HTTPS}
(if set) or \code{TRUE} (if not set)}

\item{api.base.path}{A character string specifying the base URL path to the hive}
}
\value{
A character string containing the specified API URL.
}
\description{
This function constructs a URL from the base URL of the hive and any
arguments (e.g., function names).
}
\seealso{
The function \code{\link{buildQueryString}} is used to create a query string
from the \code{query} argument.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
