% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{hiveUser-class}
\alias{hiveUser-class}
\alias{hiveUser}
\title{Class to contain data about a User}
\description{
This class is the S4 representation of the User record type.
It is a container for data about a given user.
}
\section{Slots}{

\describe{
\item{\code{username}}{A character string specifying the name of the user.
Defaults to \code{""}.}

\item{\code{password}}{A character string specifying the password of the user.
Defaults to \code{NA}.}

\item{\code{group}}{A character string specifying the name of the Group associated with
the user. Defaults to \code{""}.}

\item{\code{superuser}}{A logical value specifying whether the user is a superuser.
Defaults to \code{FALSE}.}

\item{\code{email}}{A character string specifying the email address of the user.
Defaults to \code{""}.}

\item{\code{firstName}}{A character string specifying the first name of the user.
Defaults to \code{""}.}

\item{\code{lastName}}{A character string specifying the last name of the user.
Defaults to \code{""}.}

\item{\code{dateJoined}}{A character string specifying the date that the User record was
created. Automatically populated by the hive.}

\item{\code{untrashedFileUsage}}{A numeric value specifying the file usage of the user.
Automatically populated by the hive.}

\item{\code{token}}{A character string specifying a token that can be used for password-free
authentication. Automatically populated by the hive.}

\item{\code{active}}{A logical value specifying whether the user is active.
Defaults to \code{FALSE}.}
}}

\seealso{
\describe{
  \item{group}{
    This slot corresponds to the \code{\linkS4class{hiveGroup}} object slot
    \code{name}.
  }
}
Functions for working with User records in a GeneHive are described in
\code{\link{Users}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
