% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Users.R
\name{Users}
\alias{Users}
\alias{addUser}
\alias{getUser}
\alias{updateUser}
\alias{listUsers}
\title{Add, retrieve, update, or list User record(s)}
\usage{
addUser(
  username,
  ...,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

getUser(username, con = hiveConnection())

updateUser(
  username,
  ...,
  con = hiveConnection(),
  verbose = getOption("GeneHive.verbose")
)

listUsers(con = hiveConnection(), active = TRUE)
}
\arguments{
\item{username}{A character string specifying the name of a \code{User}
record to be added, retrieved, or updated.}

\item{\dots}{Additional arguments specifying fields of the \code{User}
record to be added or updated, or fields on which to limit a listing}

\item{con}{A \code{\linkS4class{hiveConnection}} object;
if not provided, a new connection will be established}

\item{verbose}{A logical value specifying whether messages should be printed}

\item{active}{A logical value specifying whether to list only \code{User}s that are
active or inactive; a value of \code{NA} shows all \code{User}s
regardless of active status.}
}
\value{
\describe{
  \item{\code{addUser}, \code{updateUser}}{
    If the operation is successful, a \code{\linkS4class{hiveUser}} object
    object is invisibly returned.
  }
  \item{\code{getUser}}{
    If the operation is successful, a \code{\linkS4class{hiveUser}} object
    object is returned.
  }
  \item{\code{listUsers}}{
    A data frame containing one row per record and one column per field
    is returned.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add or retrieve a \code{User} record,
or to list \code{User} records.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
